+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -7,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 * Updated README.md about proposed workflow
   [#4](https://github.com/greenbone/autohooks/pull/4)
 * Allow to load plugins from a *.autohooks* directory in the git repository root
@@ -15,8 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [#8](https://github.com/greenbone/autohooks/pull/8)
 * Added a plugin API for matching files against a list of patterns
   [#9](https://github.com/greenbone/autohooks/pull/9)
+* Added plugin API for status output: ok, warning, fail and error
+  [#14](https://github.com/greenbone/autohooks/pull/14)
 
 ### Changed
+
 * Changed git repository location to https://github.com/greenbone/autohooks
 * Extend config file handling [#7](https://github.com/greenbone/autohooks/pull/7)
 * **precommit** plugin functions should accept `**kwargs` now
@@ -25,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * A Config class instance is passed as config keyword argument to the
   **precommit** plugin function. It can be used to receive plugin specific
   settings [#7](https://github.com/greenbone/autohooks/pull/7)
+* Improved check cli to verify settings and specified plugins
+  [#12](https://github.com/greenbone/autohooks/pull/12)
+* Improved and colorized terminal output [#14](https://github.com/greenbone/autohooks/pull/14)
 * Show a warning if changes created by formatting plugins conflict with stashed
   unstaged changes [#13](https://github.com/greenbone/autohooks/pull/13)